refactor: decouple setting retrieval from plugin export
